% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BeeMT}
\alias{BeeMT}
\title{honeybee database description file}
\format{
A data frame with 8 rows and 4 variables
\describe{
 \item{sampleid}{unique sample id}
 \item{Species}{scientific name}
 \item{Year}{Collection Year}
 \item{Country_abbr}{country abbr (ISO2)}
 \item{Country_Eng}{full country name}
 \item{Region_Kor}{Korean Region Name}
 \item{Region_Eng}{English Region Name}
 \item{sampleindex}{index for samples}
 \item{SubRegion2}{Subregion information 2}
 \item{SubRegion3}{Subregion information 3}
 \item{BioProject}{NCBI Bioproject}
 \item{BioSample}{NCBI BioSample}
 \item{Institute}{Institute}
 \item{Latitude}{Latitude}
 \item{Longitude}{Longitude}
 \item{Source}{Source}
 \item{Sample info}{Sample info}
 \item{status}{status}
 \item{Assembly}{annotation}
 \item{nseq}{number of sequences}
 \item{length}{sequence length}
 \item{Coding gene}{Coding gene}
 \item{tRNA}{tRNA}
 \item{rRNA}{12s rRNA}
 \item{Description}{Assembled direction}
 \item{cut_site1}{cut site}
 \item{cut_site2}{cut site2}
 \item{hapgrp}{Haplogroup}
 \item{FINAL_SAMPLEID}{FINAL_SAMPLEID}
 }
}
\usage{
BeeMT
}
\description{
A dataset for describing penguins dataset.
}
\keyword{datasets}
