% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BKtbl.R
\name{BKtbl}
\alias{BKtbl}
\title{base R's `table` extension which can receive `list` type}
\usage{
BKtbl(x, ...)
}
\arguments{
\item{x}{vector, and list objects.}
}
\description{
base R's `table` extension which can receive `list` type
}
\examples{
BKtbl(starwars$films)
}
\author{
Byungju Kim (byungju@bu.edu)
}
