% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{simple Exploratory data analysis on data frame}
\usage{
describe_df(df, desc, n_var = 7, show_freq = F, keepValues = F)
}
\arguments{
\item{desc}{column description file (under-development)}

\item{n_var}{number of values which used to distinguish factor and character.}

\item{show_freq}{calculate frequency}

\item{keepValues}{keep original numeric values for further analysis in codebook app.}

\item{x}{a data.frame or tibble object.}
}
\description{
show df's missing ratio, number of variables and column types..
}
\examples{
describe_df(billboard)
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
