% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeValueBox.R
\name{makeValuebox2}
\alias{makeValuebox2}
\title{makeValueBox2}
\usage{
makeValuebox2(x, max_height = "220px")
}
\description{
creat static value boxes for character columns.
}
\author{
byungju kim (bjk@inu.ac.kr)
}
