% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapcard.R
\name{MapCard}
\alias{MapCard}
\title{map card}
\usage{
MapCard(
  mapinfo,
  bgcolor = "#ffffff",
  fgcolor = "#696969",
  css_style = "border: 1px dashed #696969;",
  max_height = "180px"
)
}
\description{
MAP UI/Server module for card
}
\author{
bjkim@inu.ac.kr
}
