% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_theme.R
\name{new_value_box_theme}
\alias{new_value_box_theme}
\title{new_value_box_theme}
\usage{
new_value_box_theme(class = NULL, bg = NULL, fg = NULL)
}
\description{
this function exists in bslib. but somehow, it doesn't works.
}
