% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{penguins_desc}
\alias{penguins_desc}
\title{penguins data description file}
\format{
A data frame with 8 rows and 4 variables
\describe{
 \item{columns}{columns of starwars dataset columns}
 \item{desc}{short description of the column}
 \item{unit}{unit of the column}
 }
}
\usage{
penguins_desc
}
\description{
A dataset for describing penguins dataset.
}
\keyword{datasets}
