% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.df_desc.R
\name{print.df_desc}
\alias{print.df_desc}
\title{Print method for "df_desc" class}
\usage{
\method{print}{df_desc}(x, ...)
}
\arguments{
\item{x}{An object of class "df_desc"}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "df_desc" class
}
