% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRGB.R
\name{RandomRGB}
\alias{RandomRGB}
\alias{randomRGB}
\title{get random RGB code to get color}
\usage{
randomRGB(x, r = 0:255, g = 0:255, b = 0:255)
}
\arguments{
\item{x}{# of random RGB codes to print}

\item{red}{range of red color sampling [default=0:255]}

\item{green}{range of green color sampling [default=0:255]}

\item{blue}{range of blue  color sampling [default=0:255]}
}
\description{
get random RGB code to get color
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
