% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{starwars_desc}
\alias{starwars_desc}
\title{starwars data description file}
\format{
A data frame with 14 rows and 4 variables
\describe{
 \item{columns}{columns of starwars dataset columns}
 \item{desc}{short description of the column}
 \item{unit}{unit of the column}
 }
}
\usage{
starwars_desc
}
\description{
A dataset is 14 rows x 4 columns data frame which describe starwars dataframe
}
\keyword{datasets}
