% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl2df.R
\name{tbl2df}
\alias{tbl2df}
\title{Conversion of table to data frame}
\usage{
tbl2df(x)
}
\arguments{
\item{x}{a data.frame or tibble object}
}
\description{
the function converts 1D-table into data frame
}
\author{
Byungju Kim (bjk@inu.ac.kr)
}
