% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_box_theme.R
\name{value_box_theme}
\alias{value_box_theme}
\title{value_box_theme}
\usage{
value_box_theme(name = NULL, bg = NULL, fg = NULL)
}
\description{
this function exists in bslib. but somehow, it doesn't works.
}
